requirements.txt
hikari==2.0.0
hikari-lightbulb==1.2.5
beautifulsoup4==4.9.3
requests==2.25.0
urllib3==1.25.1
tabulate==0.8.7
statcord.py==3.1.0
dblpy==0.4.0
mysql-connector==2.2.9